Frameworks: BootMessage: Squashed Commits:
BootMessage: Only display package name during optimization.
The previous implementation caused the package name to show up for
other messages such as "Starting apps." or "Finishing boot.". Now
we check if the base message starts with "Optimizing" and only
display the package name if that is true.

BootMessage: Show even more info in boot dialog
Originally the boot message was: Optimizing app N of M
In http://gerrit.aicp-rom.com/#/c/1779/ this was changed to show only the package name
This change shows both of those messages on 2 lines.

Change-Id: I346129124b1167fa3b5ca9d12b617ed045b58a2f